Rename statsd.ts to statsd-utils.ts
